Moved subnet handling into Scene.cs from Wires.cs. Moved the Wire struct into it's own file. Fixed so the tools use the proper transaction api for edits.
